+++ src/ui/src/abstractlogview.cpp
@@ -2415,7 +2415,7 @@ void AbstractLogView::drawTextArea( QPaintDevice* paintDevice )
             const auto matchPart = QStringView{ logLine }.mid( match.startColumn(), match.size() );
 #endif
             const auto expandedMatchLength
-                = untabify( matchPart.toString(), expandedPrefixLength ).size();
+                = untabify( matchPart.toString(), static_cast<int>(expandedPrefixLength) ).size();
 
             const auto lengthDelta
                 = static_cast<LineLength::UnderlyingType>( expandedMatchLength - matchPart.size() );
